Move CustomConfig; change core dependency to 1.18.1
