+++ core/build.gradle
@@ -1,6 +1,6 @@
 dependencies {
     // should we use "api" instead of "implementation" anywhere?
-    implementation 'org.spigotmc:spigot:1.17.1-R0.1-SNAPSHOT'
+    implementation 'org.spigotmc:spigot:1.18.1-R0.1-SNAPSHOT'
     implementation('LibsDisguises:LibsDisguises:9.5.2') {
         exclude group: 'org.spigotmc'
         exclude group: 'com.comphenix.protocol'
