--- core/src/main/java/be/isach/ultracosmetics/api/UltraCosmeticsAPI.java
@@ -1,10 +1,0 @@
-package be.isach.ultracosmetics.api;
-
-/**
- * UltraCosmetics API.
- *
- * @author iSach
- * @since 07-23-2016
- */
-public class UltraCosmeticsAPI {
-}
